+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Add a Nocache header to the login page redirect to prevent the browser from caching the redirect page. Props [De'Yonte W.](https://github.com/rxnlabs)
 - Remove ‘password-protected’ query from redirects on successful login or logout.
+- Add favicon to password protected login page.
 
 ## [2.3] - 2020-05-17